Relax a requirement
From @mpermar:
This is interesting because the client is always open to ignore it. The problem happens when the call is sent to multiple clients. If one of them rejects it (as you are encouraging in the spec) currently the call is rejected and other clients won't be able to take it. 

So, shouldn't the spec state "CAN" instead of "SHOULD"? or perhaps as done below for hangups state something like "If a client wishes to reject a call it should send...". Otherwise, rayo servers will have to implement complex logic to handle this case.